lang_chain_practice/pyproject.toml
[project]
name = "lang-chain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]
